# packages/api-schema/openapi.json
{
  "components": {
    "schemas": {
      "AccountExport": {
        "description": "Everything Tablr holds about the caller.\n\nBacks `GET /v1/me/export`. The field list doubles as the checklist\nfor \"what is in scope for a data request\" \u2014 see the PII table in\n`backend/README.md`.",
        "properties": {
          "consents": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Consents",
            "type": "array"
          },
          "generated_at": {
            "format": "date-time",
            "title": "Generated At",
            "type": "string"
          },
          "messages": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Messages",
            "type": "array"
          },
          "profile": {
            "additionalProperties": true,
            "title": "Profile",
            "type": "object"
          },
          "receipts": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Receipts",
            "type": "array"
          },
          "reservations": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Reservations",
            "type": "array"
          },
          "tables": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Tables",
            "type": "array"
          }
        },
        "required": [
          "generated_at",
          "profile"
        ],
        "title": "AccountExport",
        "type": "object"
      },
...
      "ChatMessage": {
        "description": "One message in a table's thread.",
        "properties": {
          "body": {
            "title": "Body",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "sender_id": {
            "title": "Sender Id",
            "type": "string"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "table_id",
          "sender_id",
          "body",
          "created_at"
        ],
        "title": "ChatMessage",
        "type": "object"
      },
      "ChatPage": {
        "description": "A cursor-paginated page of messages, newest first.",
        "properties": {
          "items": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ChatMessage"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "title": "ChatPage",
        "type": "object"
      },
...
      "CreateReservationRequest": {
        "description": "Finalise a table and book it.\n\nParty size is derived from the table's members rather than sent by\nthe client, so it cannot disagree with who actually has a seat.\n`requested_for` defaults to the table's scheduled time.",
        "properties": {
          "requested_for": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested For"
          }
        },
        "title": "CreateReservationRequest",
        "type": "object"
      },
...
      "CreateTagRequest": {
        "description": "Ask for a tag that does not exist yet.",
        "properties": {
          "raw_text": {
            "maxLength": 40,
            "minLength": 2,
            "title": "Raw Text",
            "type": "string"
          }
        },
        "required": [
          "raw_text"
        ],
        "title": "CreateTagRequest",
        "type": "object"
      },
      "DeleteAccountResult": {
        "description": "Acknowledgement that deletion was accepted.\n\nAttributes:\n    scheduled: True once the soft delete has been recorded.\n    purge_after: When the hard purge runs. Null while the retention\n        window is still being decided.",
        "properties": {
          "purge_after": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purge After"
          },
          "scheduled": {
            "title": "Scheduled",
            "type": "boolean"
          }
        },
        "required": [
          "scheduled"
        ],
        "title": "DeleteAccountResult",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Me": {
        "description": "The caller's own account record.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "default": false,
            "title": "Is Verified",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "email"
        ],
        "title": "Me",
        "type": "object"
      },
      "PublicProfile": {
        "description": "What another student sees.\n\nNever includes the email address or the date of birth \u2014 only the\nage derived from it.",
        "properties": {
          "age": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Age"
          },
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "is_verified": {
            "default": false,
            "title": "Is Verified",
            "type": "boolean"
          },
          "tags": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/InterestTag"
            },
            "title": "Tags",
            "type": "array"
          }
        },
        "required": [
          "id"
        ],
        "title": "PublicProfile",
        "type": "object"
      },
      "Receipt": {
        "description": "A receipt submitted against a table.\n\nAttributes:\n    total_cents: Integer minor units, never a float.",
        "properties": {
          "created_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          },
          "currency": {
            "default": "USD",
            "title": "Currency",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "status": {
            "$ref": "#/components/schemas/ReceiptStatus",
            "default": "pending"
          },
          "submitted_by": {
            "title": "Submitted By",
            "type": "string"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          },
          "total_cents": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cents"
          }
        },
        "required": [
          "id",
          "table_id",
          "submitted_by"
        ],
        "title": "Receipt",
        "type": "object"
      },
      "ReceiptStatus": {
        "description": "Where a submitted receipt stands.",
        "enum": [
          "pending",
          "verified",
          "rejected"
        ],
        "title": "ReceiptStatus",
        "type": "string"
      },
...
      "Reservation": {
        "description": "A booking for one table.\n\nAttributes:\n    requested_for: Always UTC.\n    dispatch_channel: Which dispatcher handled it, for later\n        reconciliation.",
        "properties": {
          "dispatch_channel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispatch Channel"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "party_size": {
            "maximum": 12.0,
            "minimum": 1.0,
            "title": "Party Size",
            "type": "integer"
          },
          "requested_for": {
            "format": "date-time",
            "title": "Requested For",
            "type": "string"
          },
          "restaurant_id": {
            "title": "Restaurant Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus",
            "default": "pending"
          },
          "table_id": {
            "title": "Table Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "table_id",
          "restaurant_id",
          "party_size",
          "requested_for"
        ],
        "title": "Reservation",
        "type": "object"
      },
...
      "SendMessageRequest": {
        "description": "Post a message to a table the caller belongs to.",
        "properties": {
          "body": {
            "maxLength": 2000,
            "minLength": 1,
            "title": "Body",
            "type": "string"
          }
        },
        "required": [
          "body"
        ],
        "title": "SendMessageRequest",
        "type": "object"
      },
...
      "SubmitReceiptRequest": {
        "description": "Submit a receipt for a table the caller belongs to.\n\nAttributes:\n    image_path: Supabase Storage path. The client uploads the image\n        directly and sends only the path, so receipt photos never\n        pass through the API.",
        "properties": {
          "currency": {
            "default": "USD",
            "maxLength": 3,
            "minLength": 3,
            "title": "Currency",
            "type": "string"
          },
          "image_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Path"
          },
          "total_cents": {
            "anyOf": [
              {
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Cents"
          }
        },
        "title": "SubmitReceiptRequest",
        "type": "object"
      },
...
      "UpdateProfileRequest": {
        "description": "Partial update of the caller's own profile.\n\n`tag_ids` replaces the whole tag set when present, and leaves it\nuntouched when omitted.",
        "properties": {
          "avatar_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Avatar Url"
          },
          "bio": {
            "anyOf": [
              {
                "maxLength": 280,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bio"
          },
          "date_of_birth": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Of Birth"
          },
          "display_name": {
            "anyOf": [
              {
                "maxLength": 50,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "tag_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tag Ids"
          }
        },
        "title": "UpdateProfileRequest",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The only seam between the Expo app and Tablr's data. The app never talks to Postgres directly.",
    "title": "Tablr API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/me": {
      "delete": {
        "description": "Soft-delete the caller's account and schedule the purge.",
        "operationId": "delete_me_v1_me_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResult"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete me",
        "tags": [
          "identity"
        ]
      },
      "get": {
        "description": "Return the caller's own account record.",
        "operationId": "read_me_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get the signed-in account",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "description": "Return everything Tablr holds about the caller.",
        "operationId": "export_me_v1_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountExport"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export my data",
        "tags": [
          "identity"
        ]
      }
    },
    "/v1/me/profile": {
      "patch": {
        "description": "Apply a partial update to the caller's own profile.",
        "operationId": "update_my_profile_v1_me_profile_patch",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProfileRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update My Profile",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/v1/profiles/{profile_id}": {
      "get": {
        "description": "Return another student's public profile.",
        "operationId": "read_profile_v1_profiles__profile_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "profile_id",
            "required": true,
            "schema": {
              "title": "Profile Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PublicProfile"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Read Profile",
        "tags": [
          "profiles"
        ]
      }
    },
...
    "/v1/tables/{table_id}/messages": {
      "get": {
        "description": "Return a page of messages for this table.",
        "operationId": "list_messages_v1_tables__table_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Messages",
        "tags": [
          "chat"
        ]
      },
      "post": {
        "description": "Post a message to this table's thread.",
        "operationId": "send_message_v1_tables__table_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendMessageRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ChatMessage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send Message",
        "tags": [
          "chat"
        ]
      }
    },
    "/v1/tables/{table_id}/receipt": {
      "get": {
        "description": "Return this table's receipt. Members only.",
        "operationId": "get_receipt_v1_tables__table_id__receipt_get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Receipt",
        "tags": [
          "receipts"
        ]
      },
      "post": {
        "description": "Submit a receipt to close out this table.",
        "operationId": "submit_receipt_v1_tables__table_id__receipt_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitReceiptRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Receipt"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Submit Receipt",
        "tags": [
          "receipts"
        ]
      }
    },
    "/v1/tables/{table_id}/reservation": {
      "delete": {
        "description": "Withdraw this table's reservation. Host only.",
        "operationId": "cancel_reservation_v1_tables__table_id__reservation_delete",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Cancel Reservation",
        "tags": [
          "reservations"
        ]
      },
      "get": {
        "description": "Return this table's reservation. Members only.",
        "operationId": "get_reservation_v1_tables__table_id__reservation_get",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Reservation",
        "tags": [
          "reservations"
        ]
      },
      "post": {
        "description": "Finalise the table and book it. Host only.",
        "operationId": "create_reservation_v1_tables__table_id__reservation_post",
        "parameters": [
          {
            "in": "path",
            "name": "table_id",
            "required": true,
            "schema": {
              "title": "Table Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReservationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Reservation"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Reservation",
        "tags": [
          "reservations"
        ]
      }
    },
...
    "/v1/tags/requests": {
      "post": {
        "description": "Ask for a tag that does not exist yet.",
        "operationId": "request_tag_v1_tags_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTagRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TagRequest"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Request Tag",
        "tags": [
          "profiles"
        ]
      }
    },
...
  }
}